requirements-test.txt
boto3
pyyaml
pytest==8.3.2
pytest-cov==5.0.0